+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
 - hardis:doc:packagexml2markdown: Generate markdown documentation from a package.xml file
+- hardis:org:retrieve:packageconfig: Ignore standard Salesforce packages
 
 ## [5.3.0] 2024-10-24